requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.6
ultralytics==8.3.235
opencv-python-headless==4.8.1.78
pillow>=10.4.0
numpy>=1.26.0
torch>=2.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
mlflow==3.7.0
python-dotenv==1.0.0
requests>=2.31.0
boto3>=1.34.0